backend/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
boto3==1.26.0
PyJWT==2.6.0
gunicorn==20.1.0
requests==2.28.1
xmltodict==0.13.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
